Refactored ExternalLocationProviders
