+++ .idea/compiler.xml
@@ -36,7 +36,7 @@
       <module name="dokka.integration-tests.test" target="1.8" />
       <module name="dokka.integration.main" target="1.8" />
       <module name="dokka.integration.test" target="1.8" />
-      <module name="dokka.plugins.base.main" target="1.8" />
+      <module name="dokka.plugins.base.main" target="11" />
       <module name="dokka.plugins.base.test" target="1.8" />
       <module name="dokka.plugins.javadoc8.main" target="1.8" />
       <module name="dokka.plugins.javadoc8.test" target="1.8" />
