+++ plugins/base/src/main/kotlin/resolvers/local/DefaultLocationProviderFactory.kt
@@ -1,9 +1,10 @@
-package org.jetbrains.dokka.base.resolvers
+package org.jetbrains.dokka.base.resolvers.local
 
 import org.jetbrains.dokka.pages.RootPageNode
 import org.jetbrains.dokka.plugability.DokkaContext
 
 class DefaultLocationProviderFactory(private val context: DokkaContext) : LocationProviderFactory {
 
-    override fun getLocationProvider(pageNode: RootPageNode)  = DefaultLocationProvider(pageNode, context)
+    override fun getLocationProvider(pageNode: RootPageNode) =
+        DefaultLocationProvider(pageNode, context)
 }
